+++ lib/widgets/collection/entry_set_action_delegate.dart
@@ -22,6 +22,8 @@ import 'package:aves/widgets/common/action_mixins/size_aware.dart';
 import 'package:aves/widgets/common/extensions/build_context.dart';
 import 'package:aves/widgets/dialogs/aves_dialog.dart';
 import 'package:aves/widgets/filter_grids/album_pick.dart';
+import 'package:aves/widgets/map/map_page.dart';
+import 'package:aves/widgets/stats/stats_page.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
@@ -51,6 +53,12 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
         break;
       case CollectionAction.refreshMetadata:
         _refreshMetadata(context);
+        break;
+      case CollectionAction.map:
+        _goToMap(context);
+        break;
+      case CollectionAction.stats:
+        _goToStats(context);
         break;
       default:
         break;
@@ -240,6 +248,39 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
         // cleanup
         await storageService.deleteEmptyDirectories(selectionDirs);
       },
+    );
+  }
+
+  void _goToMap(BuildContext context) {
+    final selection = context.read<Selection<AvesEntry>>();
+    final entries = selection.isSelecting ? _getExpandedSelectedItems(selection) : context.read<CollectionLens>().sortedEntries;
+
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        settings: const RouteSettings(name: MapPage.routeName),
+        builder: (context) => MapPage(
+          entries: entries.where((entry) => entry.hasGps).toList(),
+        ),
+      ),
+    );
+  }
+
+  void _goToStats(BuildContext context) {
+    final selection = context.read<Selection<AvesEntry>>();
+    final collection = context.read<CollectionLens>();
+    final entries = selection.isSelecting ? _getExpandedSelectedItems(selection) : collection.sortedEntries.toSet();
+
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        settings: const RouteSettings(name: StatsPage.routeName),
+        builder: (context) => StatsPage(
+          entries: entries,
+          source: collection.source,
+          parentCollection: collection,
+        ),
+      ),
     );
   }
 }